This file is good to go to create RI#
This file is good to go to create RI# as a function of time at three different elevations. However, it would be interesting to fit a polynomial to the velocity gradient and then take the derivative at the tower height instead of using an Euler Forward (UF) approach, because EF causes an overestimates the Ri # from being in the denominator.